MDNS in Go POC
This is a basic rewrite of `designate-mdns` in Go. It assumes
that you're using `worker-model` style mdns, so no sending
Notifys, no polling. All it does is translate AXFR queries
to database queries, and serialize db rows into DNS protocol.
This is actually really simple, thus the small size, and single
file.

The basic features:
- Listens on TCP/UDP
- Connects to Designate's Database
- Listens for regular DNS queries (A, SOA, etc) and answers to
the best of it's ability.
- Answers AXFR queries for zones.
- Basic config via flags/file
- Makefile